Increases dpi for tesseract extraction of pdfs
Tesseract recommends a minimum dpi of 300 (https://github.com/tesseract-ocr/tesseract/wiki/ImproveQuality#rescaling) and the default pdftoppm dpi is 150. I experienced poor accuracy on some documents that was fixed by increasing the requested dpi. This could be controllable by a keyword argument, but the tesseract recommended setting seems like a better default. Love this package, it's a lifesaver.